Cleanups.

+++ tools/gen_tls_key.sh
@@ -5,19 +5,25 @@ PRIVDIR=$CERTDIR/private
 NAME=server
 SERIAL=1000
 
-if [ ! -d "$PRIVDIR" ]; then
-  mkdir -p $PRIVDIR
+if [ ! -d "$CERTDIR" ]; then
+  mkdir -p $CERTDIR
+else
+  rm -f $CERTDIR/$NAME.csr
+  rm -f $CERTDIR/$NAME.crt
+  rm -f $CERTDIR/index.txt
+  rm -f $CERTDIR/index.txt.attr
+  rm -f $CERTDIR/index.txt.old
+  rm -f $CERTDIR/$SERIAL.pem
+  rm -f $CERTDIR/serial
+  rm -f $CERTDIR/serial.old  
 fi
 
-rm -f $PRIVDIR/$NAME.key
-rm -f $CERTDIR/$NAME.csr
-rm -f $CERTDIR/$NAME.crt
-rm -f $CERTDIR/index.txt
-rm -f $CERTDIR/index.txt.attr
-rm -f $CERTDIR/index.txt.old
-rm -f $CERTDIR/$SERIAL.pem
-rm -f $CERTDIR/serial
-rm -f $CERTDIR/serial.old
+if [ ! -d "$PRIVDIR" ]; then
+  mkdir -p $PRIVDIR
+else
+  rm -f $PRIVDIR/$NAME.key
+fi
+
 
 touch $CERTDIR/index.txt
 echo $SERIAL > $CERTDIR/serial
@@ -34,7 +40,7 @@ openssl req \
   -nodes \
   -sha256 \
   -days 365 \
-  -config $CERTDIR/openssl.cnf
+  -config init/openssl.cnf
 
 openssl req -text -noout -verify -in $CERTDIR/$NAME.csr
 
@@ -48,7 +54,7 @@ openssl ca \
   -batch \
   -in $CERTDIR/$NAME.csr \
   -out $CERTDIR/$NAME.crt \
-  -config $CERTDIR/openssl.cnf \
+  -config init/openssl.cnf \
   -extensions v3_req
 
 openssl x509 -text -noout -in $CERTDIR/$NAME.crt
